WIP disable file fd handling for now.
This commit will be reset.  The code was commented out to preserve the logic
but disable the use of cache entry storage that is not used.

The problem is that the "file descriptor" is a fsal implementation specific
'thing', an fd for VFS and XFS, a pointer or data structure for ZFS and others
that use a library.  This storage is now moved to the private area of the
fsal_obj_handle.

The need for lock management has not gone away however.  At the time of this commit
there are two competing lock management methods, one in the VFS fsal and the other
in cache inode.  As with attribute and symlink management, the locks will be
maintained in cache inode and the storage will be managed by the fsal.

TODO: the locking in the fsal gets removed in deference to the locks at this
level.  The method apis change so the cache inode level can signal to the
fsal what to do with its file state (which is more than just the "fd").

Signed-off-by: Jim Lieb <jlieb@panasas.com>