NFSv4 - really respect array boundaries in bitmap4
Several functions calling nfs4_list_to_bitmap4() were
allocating memory and forgetting to free it or allocating
memory and forgetting to set the length of the buffer they've
allocated.

The cut'n'paste made this worse by spreading the code
into multiple locations, so factor out processing of 'supported attributes'
bitmap and filling the results and make sure they're used everywhere.

Also change prototype for nfs4_list_to_bitmap4() - it does not return
the counter of attributes in the list therefore it does not need
a pointer.

Signed-off-by: Max Matveev <makc@gmx.co.uk>