PROXY FSAL - restore support for 'mapped' handles
NFSv4 handle can be up-to 128 bytes long which is not going
to fit into NFSv3, let alone NFSv2, handle. Before 'New Handle'
era Proxy has had support for 'mapping' long NFSv4 handles to shorter
handles which can be used with NFSv2 and NFSv3, 'New Handle' made
this trick impossible. Now with New API we can do it again: use
16 bytes for encode fileid and hash of the NFSv4 handle and give it
to callers who ask for NFSv2 or NFSv3 digest. First 2 bytes of the
'mapped' handle are used to disambiguate it from full NFSv4 handle.

On the database side the mapping is from object+hash to the full
NFSv4 handle from the backend server. Because NFSv4 handle is
variable in lenght extra care is needed then loading those handles
from the database.

Also there is no more libhandle_mapping - the code is linked into
the single libfsalproxy DSO. Dependency on sqlite3 is on the libfsalproxy,
not on the target ganesha binary.

Signed-off-by: Max Matveev <makc@gmx.co.uk>