PROXY FSAL - use single socket to talk to backend
Instead of single RPC client shared across all threads or per-worker
RPC client as it was done in the previous implementation of Proxy's RPC
communication use single socket which can be used to talk to the backend
NFSv4 server.

There is a thread which deals with maintaining the connection, reading
replies from the socket and matching them to the XIDs which had been
sent out.

Each NFS worker thread which sends data out needs to get a 'context'
which contains buffers to encode/decode data from plus some locks and
conditional variables to wait on.  There is a fixed number of those
contexts and it is used for both controlling number of concurrent
requests to the backend and to avoid constant churn of creating and
destroying mutexes.

The 'sending' thread 'donates' a buffer which is used by the receiving
thread to dump reply data into when receiver finds an XID match. The
decoding of the reply is done by the sending thread, not the receiver.

The old 'client id renewer' thread uses the same mechanism to talk
to backend NFS server.

This change removes support for UDP.

Signed-off-by: Max Matveev <makc@gmx.co.uk>