PROXY FSAL - use per-thread context for IO
PROXY FSAL used to switch operation context for read and write to the
context that was used when the file was open. With cache inode rewrite
multiple IO requests for the same file can be executed in parallel which
makes for an unsafe use of that context: rpc client inside the context
cannot be used from multiple threads at the same time.

In order to keep multithreding we now use the "natural" context of
the worker.

Signed-off-by: Max Matveev <makc@gmx.co.uk>